Extended functionality of webView.
Added history back/forth buttons, as well as a vote button so the user
does not have to return to the feed view to vote on a post. Now I must
figure out a way to access the item renderer for the feed view so that
the vote button there is updated when manipulated in the webView. Then
should be ready for release!